+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ### v0.14.6
 
+* Add support for Xcode Cloud.
 * Add support for Xcode Server.
 
 ### v0.14.5 (2022-04-22)